requirements.txt
langgraph>=0.6.6
langchain-google-genai>=2.1.9
langchain>=0.3.0
python-dotenv>=1.1.1
requests>=2.32.5
langsmith>=0.4.16
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
bm25s>=0.2.0